fixed testing dependency
