+++ src/test/kotlin/org/wfanet/measurement/eventdataprovider/privacybudgetmanagement/BUILD.bazel
@@ -40,6 +40,7 @@ kt_jvm_test(
     test_class = "org.wfanet.measurement.eventdataprovider.privacybudgetmanagement.PrivacyBucketFilterTest",
     deps = [
         "//src/main/kotlin/org/wfanet/measurement/eventdataprovider/privacybudgetmanagement:privacy_budget_manager",
+        "//src/main/kotlin/org/wfanet/measurement/eventdataprovider/privacybudgetmanagement/testing",
         "//src/main/proto/wfa/measurement/api/v2alpha:measurement_spec_kt_jvm_proto",
         "//src/main/proto/wfa/measurement/api/v2alpha:requisition_spec_kt_jvm_proto",
         "//src/main/proto/wfa/measurement/api/v2alpha/event_templates/testing:event_kt_jvm_proto",
@@ -60,6 +61,7 @@ kt_jvm_test(
     test_class = "org.wfanet.measurement.eventdataprovider.privacybudgetmanagement.PrivacyBudgetManagerTest",
     deps = [
         "//src/main/kotlin/org/wfanet/measurement/eventdataprovider/privacybudgetmanagement:privacy_budget_manager",
+        "//src/main/kotlin/org/wfanet/measurement/eventdataprovider/privacybudgetmanagement/testing",
         "@wfa_common_jvm//imports/java/com/google/common/truth",
         "@wfa_common_jvm//imports/java/org/junit",
         "@wfa_common_jvm//imports/kotlin/kotlin/test",
